deg2rad is a trigonometric function
