+++ ivy/functional/frontends/numpy/mathematical_functions/trigonometric_functions.py
@@ -164,3 +164,25 @@ def cosh(
 
 
 cosh.unsupported_dtypes = {"torch": ("float16",)}
+
+
+@from_zero_dim_arrays_to_float
+def deg2rad(
+    x,
+    /,
+    out=None,
+    *,
+    where=True,
+    casting="same_kind",
+    order="K",
+    dtype=None,
+    subok=True,
+    signature=None,
+    extobj=None,
+):
+    if dtype:
+        x = ivy.astype(ivy.array(x), ivy.as_ivy_dtype(dtype))
+    ret = ivy.multiply(ivy.divide(x, 180), ivy.pi, out=out)
+    if ivy.is_array(where):
+        ret = ivy.where(where, ret, ivy.default(out, ivy.zeros_like(ret)), out=out)
+    return ret
